% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eif-package.R
\docType{package}
\name{eif-package}
\alias{eif}
\alias{eif-package}
\title{eif: S3 Vector Class for and Estimand and It's Efficient Influene Function}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Nicholas Williams \email{ntwilliams.personal@gmail.com} (\href{https://orcid.org/0000-0002-1378-4831}{ORCID}) [copyright holder]

}
\keyword{internal}
